# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Starter API",
    "contact": {}
  },
  "host": "localhost:8080.",
  "paths": {
...
    "/api/v1/creativeSpace": {
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Создание креативной площадки",
        "parameters": [
          {
            "description": "Параметры запроса",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requests.createCreativeSpaceRequestData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.createCreativeSpaceResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/creativeSpaces": {
      "get": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Возвращает список креативных площадок",
        "parameters": [
          {
            "type": "string",
            "description": "Authentication header",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.getCreativeSpacesResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/creativeSpaces/:id": {
      "get": {
        "produces": ["application/json"],
        "summary": "Возвращает информацию о пользователе",
        "parameters": [
          {
            "type": "string",
            "description": "Id креативной площадки",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.getCreativeSpaceResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      },
      "delete": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Удаление информации о креативной площадке",
        "parameters": [
          {
            "type": "string",
            "description": "Id креативной площадки",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      },
      "patch": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Обновление информации о креативном пространстве",
        "parameters": [
          {
            "type": "string",
            "description": "Id креативного пространства",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Параметры запроса",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requests.patchCreativeSpaceRequestData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
...
    "/api/v1/session": {
      "get": {
        "summary": "Получение сессии",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      },
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Создание сессии пользователя",
        "parameters": [
          {
            "description": "Параметры запроса",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requests.createSessionRequestData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.createSessionResponseData"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "delete": {
        "summary": "Удаление сессии",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "model.ResponseError": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/model.ResponseErrorField"
        },
        "success": {
          "type": "boolean",
          "default": false
        }
      }
    },
    "model.ResponseErrorField": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "details": {
          "type": "string"
        }
      }
    },
    "model.ResponseSuccess": {
      "type": "object",
      "properties": {
        "data": {},
        "success": {
          "type": "boolean",
          "default": true
        }
      }
    },
...
    "requests.createCreativeSpaceRequestCalendar": {
      "type": "object",
      "required": ["events", "workDayIndexes"],
      "properties": {
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/requests.createCreativeSpaceRequestCalendarEvent"
          }
        },
        "link": {
          "type": "string"
        },
        "workDayIndexes": {
          "type": "array",
          "items": {
            "type": "integer"
          }
        }
      }
    },
    "requests.createCreativeSpaceRequestCalendarEvent": {
      "type": "object",
      "required": ["date"],
      "properties": {
        "date": {
          "type": "string"
        }
      }
    },
    "requests.createCreativeSpaceRequestCoordinate": {
      "type": "object",
      "required": ["latitude", "longitude"],
      "properties": {
        "latitude": {
          "type": "number"
        },
        "longitude": {
          "type": "number"
        }
      }
    },
    "requests.createCreativeSpaceRequestData": {
      "type": "object",
      "required": [
        "address",
        "area",
        "calendar",
        "capacity",
        "coordinate",
        "description",
        "photos",
        "pricePerDay",
        "spaceType",
        "title"
      ],
      "properties": {
        "address": {
          "type": "string"
        },
        "area": {
          "type": "integer"
        },
        "calendar": {
          "$ref": "#/definitions/requests.createCreativeSpaceRequestCalendar"
        },
        "capacity": {
          "type": "integer"
        },
        "coordinate": {
          "$ref": "#/definitions/requests.createCreativeSpaceRequestCoordinate"
        },
        "description": {
          "type": "string"
        },
        "metroStations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/requests.createCreativeSpaceRequestMetroStation"
          }
        },
        "photos": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "pricePerDay": {
          "type": "integer"
        },
        "spaceType": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "requests.createCreativeSpaceRequestMetroStation": {
      "type": "object",
      "required": ["distanceInMinutes", "id"],
      "properties": {
        "distanceInMinutes": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        }
      }
    },
...
    "requests.createSessionRequestData": {
      "type": "object",
      "required": ["password", "phone"],
      "properties": {
        "password": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      }
    },
...
    "requests.patchCreativeSpaceRequestCalendar": {
      "type": "object",
      "properties": {
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/requests.patchCreativeSpaceRequestCalendarEvent"
          }
        },
        "link": {
          "type": "string"
        },
        "workDayIndexes": {
          "type": "array",
          "items": {
            "type": "integer"
          }
        }
      }
    },
    "requests.patchCreativeSpaceRequestCalendarEvent": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string"
        }
      }
    },
    "requests.patchCreativeSpaceRequestCoordinate": {
      "type": "object",
      "properties": {
        "latitude": {
          "type": "number"
        },
        "longitude": {
          "type": "number"
        }
      }
    },
    "requests.patchCreativeSpaceRequestData": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "area": {
          "type": "integer"
        },
        "calendar": {
          "$ref": "#/definitions/requests.patchCreativeSpaceRequestCalendar"
        },
        "capacity": {
          "type": "integer"
        },
        "coordinate": {
          "$ref": "#/definitions/requests.patchCreativeSpaceRequestCoordinate"
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "metroStations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/requests.patchCreativeSpaceRequestMetroStation"
          }
        },
        "photos": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "pricePerDay": {
          "type": "integer"
        },
        "spaceType": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "requests.patchCreativeSpaceRequestMetroStation": {
      "type": "object",
      "properties": {
        "distanceInMinutes": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        }
      }
    },
...
    "responses.createCreativeSpaceResponseCreativeSpace": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        }
      }
    },
    "responses.createCreativeSpaceResponseData": {
      "type": "object",
      "properties": {
        "creativeSpace": {
          "$ref": "#/definitions/responses.createCreativeSpaceResponseCreativeSpace"
        }
      }
    },
...
    "responses.createSessionResponseData": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/responses.createSessionResponseUser"
        }
      }
    },
    "responses.createSessionResponseUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "phone": {
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      }
    },
...
    "responses.getCreativeSpaceResponseCalendar": {
      "type": "object",
      "properties": {
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.getCreativeSpaceResponseCalendarEvent"
          }
        },
        "link": {
          "type": "string"
        },
        "workDayIndexes": {
          "type": "array",
          "items": {
            "type": "integer"
          }
        }
      }
    },
    "responses.getCreativeSpaceResponseCalendarEvent": {
      "type": "object",
      "properties": {
        "bookingId": {
          "type": "integer"
        },
        "date": {
          "type": "string"
        }
      }
    },
    "responses.getCreativeSpaceResponseCoordinate": {
      "type": "object",
      "properties": {
        "latitude": {
          "type": "number"
        },
        "longitude": {
          "type": "number"
        }
      }
    },
    "responses.getCreativeSpaceResponseCreativeSpace": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "area": {
          "type": "integer"
        },
        "averageRating": {
          "type": "integer"
        },
        "calendar": {
          "$ref": "#/definitions/responses.getCreativeSpaceResponseCalendar"
        },
        "capacity": {
          "type": "integer"
        },
        "coordinate": {
          "$ref": "#/definitions/responses.getCreativeSpaceResponseCoordinate"
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "landlordInfo": {
          "$ref": "#/definitions/responses.getCreativeSpaceResponseLandlordInfo"
        },
        "metroStations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.getCreativeSpaceResponseMetroStation"
          }
        },
        "photos": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "pricePerDay": {
          "type": "integer"
        },
        "scores": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.getCreativeSpaceResponseScore"
          }
        },
        "spaceType": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "responses.getCreativeSpaceResponseData": {
      "type": "object",
      "properties": {
        "creativeSpace": {
          "$ref": "#/definitions/responses.getCreativeSpaceResponseCreativeSpace"
        }
      }
    },
    "responses.getCreativeSpaceResponseLandlordInfo": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "inn": {
          "type": "string"
        },
        "legalEntityName": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        }
      }
    },
    "responses.getCreativeSpaceResponseMetroStation": {
      "type": "object",
      "properties": {
        "color": {
          "type": "string"
        },
        "distanceInMinutes": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "responses.getCreativeSpaceResponseScore": {
      "type": "object",
      "properties": {
        "comment": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "rating": {
          "type": "integer"
        },
        "user": {
          "$ref": "#/definitions/responses.getCreativeSpaceResponseScoreUser"
        }
      }
    },
    "responses.getCreativeSpaceResponseScoreUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        }
      }
    },
    "responses.getCreativeSpacesResponseCalendar": {
      "type": "object",
      "properties": {
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.getCreativeSpacesResponseCalendarEvent"
          }
        },
        "link": {
          "type": "string"
        },
        "workDayIndexes": {
          "type": "array",
          "items": {
            "type": "integer"
          }
        }
      }
    },
    "responses.getCreativeSpacesResponseCalendarEvent": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string"
        }
      }
    },
    "responses.getCreativeSpacesResponseCoordinate": {
      "type": "object",
      "properties": {
        "latitude": {
          "type": "number"
        },
        "longitude": {
          "type": "number"
        }
      }
    },
    "responses.getCreativeSpacesResponseCreativeSpace": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "area": {
          "type": "integer"
        },
        "averageRating": {
          "type": "integer"
        },
        "calendar": {
          "$ref": "#/definitions/responses.getCreativeSpacesResponseCalendar"
        },
        "capacity": {
          "type": "integer"
        },
        "coordinate": {
          "$ref": "#/definitions/responses.getCreativeSpacesResponseCoordinate"
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "landlordId": {
          "type": "integer"
        },
        "metroStations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.getCreativeSpacesResponseMetroStation"
          }
        },
        "photos": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "pricePerDay": {
          "type": "integer"
        },
        "scores": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.getCreativeSpacesResponseScore"
          }
        },
        "spaceType": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "responses.getCreativeSpacesResponseData": {
      "type": "object",
      "properties": {
        "creativeSpaces": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.getCreativeSpacesResponseCreativeSpace"
          }
        }
      }
    },
    "responses.getCreativeSpacesResponseMetroStation": {
      "type": "object",
      "properties": {
        "color": {
          "type": "string"
        },
        "distanceInMinutes": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "responses.getCreativeSpacesResponseScore": {
      "type": "object",
      "properties": {
        "comment": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "rating": {
          "type": "integer"
        },
        "user": {
          "$ref": "#/definitions/responses.getCreativeSpacesResponseScoreUser"
        }
      }
    },
    "responses.getCreativeSpacesResponseScoreUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        }
      }
    },
...
  }
}